eml2xlsx/requirements_eml2xlsx.txt
openpyxl
beautifulsoup4
lxml
python-dateutil
colorama
